Add files via upload
I have created a HTML page to perform the functions. I have used Visual Code Studio for this. 

There are two ways of extracting the data from datasheet in the HTML document. First method is installing Live Server extension in VS code and then we can run the HTML files with live reloading and see changes instantly. Second method is to manually open the browser. In this method the major challenge is security restrictions.
I have used the second method, hence to avoid the restrictions I have used FileReader API to read the files and Papa Parse to convert the csv data into usable format.

Output is in the file.